Exer 13.9-12
13. 9 Sequelize provides a set of pre-defined validations for the model fields, which it performs before storing the objects in the database.

It's decided to change the user creation policy so that only a valid email address is valid as a username. Implement validation that verifies this issue during the creation of a user.

Modify the error handling middleware to provide a more descriptive error message of the situation (for example, using the Sequelize error message), e.g.

13.10. Expand the application so that the current logged-in user identified by a token is linked to each blog added. To do this you will also need to implement a login endpoint POST /api/login, which returns the token.

13.11 Make deletion of a blog only possible for the user who added the blog.

13.12 Modify the routes for retrieving all blogs and all users so that each blog shows the user who added it and each user shows the blogs they have added.