Exer 13.13
13.13
Implement filtering by keyword in the application for the route returning all blogs. The filtering should work as follows

GET /api/blogs?search=react returns all blogs with the search word react in the title field, the search word is case-insensitive
GET /api/blogs returns all blogs
This should be useful for this task and the next one.