Exer 13.24
13.24 Grand finale: towards the end of part 4 there was mention of a token-criticality problem: if a user's access to the system is decided to be revoked, the user may still use the token in possession to use the system.

The usual solution to this is to store a record of each token issued to the client in the backend database, and to check with each request whether access is still valid. In this case, the validity of the token can be removed immediately if necessary. Such a solution is often referred to as a server-side session.

Now expand the system so that the user who has lost access will not be able to perform any actions that require login.

You will probably need at least the following for the implementation

a boolean value column in the user table to indicate whether the user is disabled

it is sufficient to disable and enable users directly from the database
a table that stores active sessions

a session is stored in the table when a user logs in, i.e. operation POST /api/login
the existence (and validity) of the session is always checked when the user makes an operation that requires login
a route that allows the user to "log out" of the system, i.e. to practically remove active sessions from the database, the route can be e.g. DELETE /api/logout
Keep in mind that actions requiring login should not be successful with an "expired token", i.e. with the same token after logging out